# backend/.well-known/openapi.yaml
openapi: 3.0.0
info:
  title: InsightFlow MCP Server API
  description: MCP (Model Context Protocol) server for managing projects, stacks, roadmaps, tasks, and subtasks
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://your-domain.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase authentication token
paths:
...
  /mcp/stacks/get:
    post:
      summary: Get stack
      description: Get a single stack by ID
      operationId: getStack
      tags:
        - Stacks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stack_id
              properties:
                stack_id:
                  type: string
      responses:
        '200':
          description: Stack details
  /mcp/stacks/create:
    post:
      summary: Create stack
      description: Create a new stack
      operationId: createStack
      tags:
        - Stacks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project_id
                - name
              properties:
                project_id:
                  type: string
                name:
                  type: string
                technology:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Stack created
  /mcp/stacks/update:
    post:
      summary: Update stack
      description: Update an existing stack
      operationId: updateStack
      tags:
        - Stacks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stack_id
              properties:
                stack_id:
                  type: string
                name:
                  type: string
                technology:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Stack updated
  /mcp/stacks/delete:
    post:
      summary: Delete stack
      description: Delete a stack
      operationId: deleteStack
      tags:
        - Stacks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stack_id
              properties:
                stack_id:
                  type: string
      responses:
        '200':
          description: Stack deleted
...
  /mcp/roadmaps/get:
    post:
      summary: Get roadmap
      description: Get a single roadmap with tasks and subtasks
      operationId: getRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
              properties:
                roadmap_id:
                  type: string
      responses:
        '200':
          description: Roadmap details
  /mcp/roadmaps/create:
    post:
      summary: Create roadmap
      description: Create a new roadmap
      operationId: createRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project_id
                - name
              properties:
                project_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [draft, active, completed]
      responses:
        '201':
          description: Roadmap created
  /mcp/roadmaps/update:
    post:
      summary: Update roadmap
      description: Update an existing roadmap
      operationId: updateRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
              properties:
                roadmap_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Roadmap updated
  /mcp/roadmaps/delete:
    post:
      summary: Delete roadmap
      description: Delete a roadmap
      operationId: deleteRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
              properties:
                roadmap_id:
                  type: string
      responses:
        '200':
          description: Roadmap deleted
...
  /mcp/subtasks/get:
    post:
      summary: Get subtask
      description: Get a single subtask
      operationId: getSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subtask_id
              properties:
                subtask_id:
                  type: string
      responses:
        '200':
          description: Subtask details
  /mcp/subtasks/create:
    post:
      summary: Create subtask
      description: Create a new subtask
      operationId: createSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
                - name
              properties:
                task_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        '201':
          description: Subtask created
  /mcp/subtasks/update:
    post:
      summary: Update subtask
      description: Update an existing subtask
      operationId: updateSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subtask_id
              properties:
                subtask_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        '200':
          description: Subtask updated
  /mcp/subtasks/delete:
    post:
      summary: Delete subtask
      description: Delete a subtask
      operationId: deleteSubtask
      tags:
        - Subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subtask_id
              properties:
                subtask_id:
                  type: string
      responses:
        '200':
          description: Subtask deleted
...